+++ lib/govuk_app_config/govuk_healthcheck/redis.rb
@@ -1,0 +1,21 @@
+module GovukHealthcheck
+  class Redis
+    def name
+      :redis
+    end
+
+    def status
+      client = ::Redis.new
+
+      client.set("healthcheck", "val")
+      client.get("healthcheck")
+      client.del("healthcheck")
+
+      client.close
+
+      GovukHealthcheck::OK
+    rescue StandardError
+      GovukHealthcheck::CRITICAL
+    end
+  end
+end